ALTERAÇÕES FINAIS BLOG PESSOAL 22/11/2021

+++ blogPessoal/blogPessoal/src/main/java/ogr/generation/blogPessoal/Controller/PostagemController.java
@@ -1,6 +1,9 @@
 package ogr.generation.blogPessoal.Controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,14 +16,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import ogr.generation.blogPessoal.model.Postagem;
 import ogr.generation.blogPessoal.repository.PostagemRepository;
 
 @RestController
 @RequestMapping("/postagens")
-@CrossOrigin("*")
+@CrossOrigin(value = "*", allowedHeaders = "*")
 
 public class PostagemController {
 
@@ -46,17 +51,23 @@ public class PostagemController {
 	}
 	
 	@PostMapping
-	public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
+	public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
 		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
 	}
 	
 	@PutMapping
-	public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+	public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
+		return repository.findById(postagem.getId())
+		        .map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+		        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
 	public void delete(@PathVariable long id) {
-		repository.deleteById(id);
+		Optional<Postagem> post = repository.findById(id);
+        if(post.isEmpty())
+	        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        repository.deleteById(id);
 	}
 }
